requirements.txt
pyyaml==5.4.1
python-miio==0.5.4